review 6 (delta-scoped): five fixes in the review-round code itself
The sixth pass reviewed only the code written DURING the five review rounds
(804ced7..master). Every finding was independently verified against a live
repro before any fix, and the full batteries re-ran after (landing CPU pins,
NF4 GPU ring, real-checkpoint TE parity - all green).

- embedder.py vision rung: the small-total-RAM refusal never consulted
  FIZGIG_NO_TE_H2D - with streaming killed, a refusal about the STREAMED
  build's RAM staging fired on a path that would have loaded resident.
  It now stands down when the kill-switch is set.
- embedder.py vision rung: the ~19 GB RAM heads-up printed above the gate
  that decides whether streaming happens at all; moved inside the branch
  that actually streams.
- embedder.py text rung: the resident-ram-short message promised 'slower,
  works' on exactly the tier (12 GB card, 16 GB RAM) where the resident
  build cannot fit; now honest wording, and plan_text_te_build grew a
  ram_ok input wired to FIZGIG_TE_RAM_OK - the same override the vision
  rung already honours - so a transiently-low RAM reading can be overridden.
- model.py forward_cached: the dtype-guard sweep missed the video patch-proj
  twin (its own audio sibling five lines down WAS guarded) - the bf16
  checkpoint forwarded but still died in forward_cached. Mirrors forward()
  exactly again; seeded repro confirms bitwise-identical outputs.
- h3_nf4_h2d_offload.py: the pin-disarm margin was a flat +10 GB, so a
  ~1.6 GB NF4 stage lost pinning (and the ring's prefetch overlap) on any
  box under ~11.6 GB available. Margin now scales (0.75x est, 4 GB floor),
  preserving ~23 GB at the 13 GB bf16 cap the guard was written for. The
  trainer's gate log no longer blames the 'OS pin limit' for what may be
  a policy disarm.

+++ src/fizgig/minimax/model.py
@@ -885,7 +885,8 @@ class MiniMaxH3DiT(nn.Module):
             if text_states.shape[-1] != self.hidden_size:
                 text_states = self.token_refiner(self.condition_proj(text_states))
             video_rows = patchify_video(video_latent.to(torch.float32), self.patch_size)
-            video_embed = self.video_patch_proj(video_rows).to(dtype)
+            video_embed = self.video_patch_proj(
+                video_rows.to(self.video_patch_proj.weight.dtype)).to(dtype)
 
             t_val = t.reshape(-1)[:1].to(torch.float32) if torch.is_tensor(t) else torch.tensor([float(t)], device=device)
             t_val = t_val.to(device)
